app/requirements.txt
Flask
gunicorn
pymongo[tls,srv]
python-dotenv